refactor(core): Move `IdentifierSearchRequest` and `IdentifierSearchResult`
