+++ src/core/model/search.rs
@@ -1,5 +1,35 @@
 use crate::core::lookup::lookup_table::LookupTableLevel;
+use crate::core::model::direction::Direction;
 use crate::core::Identifier;
+
+#[derive(Debug, Clone, Copy)]
+pub struct IdentifierSearchRequest {
+    pub target: Identifier,
+    pub level: LookupTableLevel,
+    pub direction: Direction,
+}
+
+impl IdentifierSearchRequest {
+    pub fn new(target: Identifier, level: LookupTableLevel, direction: Direction) -> Self {
+        IdentifierSearchRequest {
+            target,
+            level,
+            direction,
+        }
+    }
+
+    pub fn target(&self) -> &Identifier {
+        &self.target
+    }
+
+    pub fn level(&self) -> LookupTableLevel {
+        self.level
+    }
+
+    pub fn direction(&self) -> Direction {
+        self.direction
+    }
+}
 
 /// A struct representing the result of an identifier search within lookup table of current node.
 ///
